Update organization offices users controller and policy so it checks for office in context

+++ app/policies/organization/offices/user_policy.rb
@@ -3,6 +3,8 @@
 module Organization
   module Offices
     class UserPolicy < Organization::UserPolicy
+      authorize :office, optional: true
+
       alias_rule :show?, :edit?, :update?, :undiscard?, :undiscard_all?, to: :not_supported?
       alias_rule :index?, :new?, :create?, to: :manage?
       alias_rule :remove_all?, :destroy_all?, to: :manage?
@@ -11,10 +13,17 @@ module Organization
 
       def manage?
         if record == User
-          ddfip_admin? || supervisor?
+          ddfip_admin? || supervisor_of_office?
         elsif record.is_a? User
-          (ddfip_admin? || supervisor?) && organization_match?(record)
+          organization_match?(record) &&
+            (ddfip_admin? || (supervisor_of_office? && supervisor_of?(record)))
         end
+      end
+
+      private
+
+      def supervisor_of_office?
+        supervisor? && (office.nil? || supervisor_of?(office))
       end
     end
   end
